Split generate mocks code
